Factor out common implementations and remove unused code
